requirements.txt
Django>=2.2.5
django-cryptography>=0.3
django-encrypted-filefield>=0.2.1
django-file-md5>=1.0.3
django-formtools>=2.1
django-otp>=0.7.1
django-otp-yubikey>=0.5.2
django-phonenumber-field>=1.3.0
django-random-filestorage>=0.1.0
django-secure>=1.0.1
django-sslserver>=0.21
django-two-factor-auth>=1.9.1
djangorestframework>=3.10.3
djangorestframework-api-key>=1.4.1
Pillow>=6.1.0
python-magic>=0.4.15
django-session-security